requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
black==23.3.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
python-dotenv==1.0.0
tomli==2.0.1
yarl==1.8.2
